move delete course confirm to course-list component

+++ src/app/courses/course-list/course-list.component.ts
@@ -39,15 +39,17 @@ export class CourseListComponent implements OnInit {
   };
   
   onCourseDeleted(courseId: number) {
-    const courseIdToServer = courseId.toString();
-    this.coursesService.removeCourse(courseIdToServer)
-      .subscribe(
-        () => {
-          this.callCourses();
-        },
-        (error) => console.log(error)
-      );
-      
+    const courseToBeDeleted = confirm("Do you really want to delete this course?");
+    if (courseToBeDeleted) {
+      const courseIdToServer = courseId.toString();
+      this.coursesService.removeCourse(courseIdToServer)
+        .subscribe(
+          () => {
+            this.callCourses();
+          },
+          (error) => console.log(error)
+        );
+    }
   }
 
   onLoadMore() {
